Turn AbstractContentValues into a generic class.
Summary:
In order to make the process of transform a POJO, created by Retrofit, into a
ContentValue, an abstract method method with a generic type was added to the
AbstractContentValues definition.

Thus any subclass, which is also an adapter to a SQLite table, will have to
tie the POJO to the concrete implementation, turning the SyncAdapter loosely
coupled to content provider implementation.